+++ src/test/java/com/github/sergueik/selenium/PrintToPDFDevToolsTest.java
@@ -1,6 +1,6 @@
 package com.github.sergueik.selenium;
 /**
- * Copyright 2022 Serguei Kouzmine
+ * Copyright 2022,2023 Serguei Kouzmine
  */
 
 import static org.hamcrest.CoreMatchers.containsString;
@@ -21,6 +21,8 @@ import java.io.UnsupportedEncodingException;
 import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Paths;
+import java.text.NumberFormat;
+import java.text.ParseException;
 import java.util.Optional;
 
 import org.junit.After;
@@ -54,11 +56,16 @@ import org.openqa.selenium.devtools.v113.page.Page.PrintToPDFResponse;
  * Selected test scenarios for Selenium 4 Chrome Developer Tools bridge
  * 
  * https://chromedevtools.github.io/devtools-protocol/tot/Page/#method-printToPDF
+ * for page dimensions see https://papersizes.io/a/a4
+ * https://css.paperplaza.net/conferences/support/page.php
+ * see also:
+ * https://stackoverflow.com/questions/20904191/pdfbox-find-page-dimensions
  */
 // NOTE: not extending BaseDevToolsTest
 
 public class PrintToPDFDevToolsTest {
 
+	private static boolean keepFile = true;
 	private static String magic = null;
 	private static boolean runHeadless = true;
 	private static PrintToPDFResponse response;
@@ -118,7 +125,8 @@ public class PrintToPDFDevToolsTest {
 
 	@After
 	public void afterTest() {
-		new File(System.getProperty("user.dir") + "/" + filename).delete();
+		if (!keepFile)
+			new File(System.getProperty("user.dir") + "/" + filename).delete();
 		driver.get("about:blank");
 	}
 
@@ -152,7 +160,7 @@ public class PrintToPDFDevToolsTest {
 			magic = body.substring(0, 9);
 			assertThat(magic, containsString("%PDF"));
 
-			filename = "result2.pdf";
+			filename = "result3.pdf";
 			writeToFile(Base64.decodeBase64(response.getData().getBytes("UTF8")),
 					filename);
 		} catch (UnsupportedEncodingException e) {
@@ -161,24 +169,9 @@ public class PrintToPDFDevToolsTest {
 
 	}
 
-	/*
-	 * response = chromeDevTools.send(new
-	 * Command<PrintToPDFResponse>("Page.printToPDF",
-	 * ImmutableMap.of("landscape", landscape),
-	 * ImmutableMap.of("displayHeaderFooter", displayHeaderFooter),
-	 * ImmutableMap.of("printBackground", printBackground),
-	 * ImmutableMap.of("scale", scale), ImmutableMap.of("paperWidth", null),
-	 * ImmutableMap.of("paperWidth", null), ImmutableMap.of("marginTop", null),
-	 * ImmutableMap.of("marginBottom", null), ImmutableMap.of("marginLeft",
-	 * null), ImmutableMap.of("marginRight", null),
-	 * ImmutableMap.of("pageRanges", null),
-	 * ImmutableMap.of("headerTemplate", null),
-	 * ImmutableMap.of("footerTemplate", null),
-	 * ImmutableMap.of("preferCSSPageSize", 0), ImmutableMap.of("transferMode",
-	 * transferMode), ConverterFunctions.map("data",
-	 * PrintToPDFResponse.class)));
-	 */
-
+	// verify paper size is Letter
+	// NOTE: top margin is different on the first page and the rest
+	// (not attempted in this test)
 	@Test
 	public void test3() {
 		// Act
@@ -193,7 +186,7 @@ public class PrintToPDFDevToolsTest {
 			assertThat(body, notNullValue());
 			magic = body.substring(0, 9);
 			assertThat(magic, containsString("%PDF"));
-			filename = "result3.pdf";
+			filename = "result4.pdf";
 			writeToFile(Base64.decodeBase64(response.getData().getBytes("UTF8")),
 					filename);
 
@@ -205,6 +198,8 @@ public class PrintToPDFDevToolsTest {
 			assertThat(pdf.text, containsString("Français"));
 			assertThat(pdf.encrypted, is(false));
 			assertThat(pdf.numberOfPages, equalTo(2));
+			assertThat(pdf.getHeight(), equalTo(11.0));
+			assertThat(pdf.getWidth(), equalTo(8.5));
 
 		} catch (UnsupportedEncodingException e) {
 			System.err.println("Exception (ignored): " + e.toString());
@@ -224,6 +219,7 @@ public class PrintToPDFDevToolsTest {
 	@Test
 	public void test4() {
 		// Act
+		filename = "result5.pdf";
 		response = chromeDevTools.send(new Command<PrintToPDFResponse>(
 				"Page.printToPDF", ImmutableMap.of("landscape", landscape),
 				ConverterFunctions.map("data", PrintToPDFResponse.class)));
@@ -234,7 +230,6 @@ public class PrintToPDFDevToolsTest {
 			assertThat(body, notNullValue());
 			magic = body.substring(0, 9);
 			assertThat(magic, containsString("%PDF"));
-			filename = "result4.pdf";
 			writeToFile(Base64.decodeBase64(response.getData().getBytes("UTF8")),
 					filename);
 
@@ -243,6 +238,76 @@ public class PrintToPDFDevToolsTest {
 		}
 
 	}
+
+	// print A4 (default is Letter)
+	// for page dimensions see https://papersizes.io/a/a4
+	// https://css.paperplaza.net/conferences/support/page.php
+	// NOTE: top margin is different on the first page and the rest
+	// (not attempted in this test)
+	@Test
+	public void test5() {
+		// Act
+		response = chromeDevTools.send(new Command<PrintToPDFResponse>(
+				"Page.printToPDF", ImmutableMap.of("landscape", landscape),
+				o -> o.read(PrintToPDFResponse.class)));
+		filename = "result6.pdf";
+		response = chromeDevTools.send(Page.printToPDF(Optional.of(landscape),
+				Optional.of(displayHeaderFooter), Optional.of(printBackground),
+				Optional.of(scale), Optional.of(8.27), Optional.of(11.69),
+				Optional.of(1.0), Optional.of(1.44), Optional.of(0.75),
+				Optional.of(0.52), Optional.empty(/* pageRanges */),
+				Optional.empty(/* headerTemplate */),
+				Optional.empty(/* footerTemplate */),
+				Optional.of(true /* preferCSSPageSize */), Optional.of(
+						transferMode /* Allowed Values: ReturnAsBase64, ReturnAsStream */ )));
+
+		assertThat(response, notNullValue());
+
+		try {
+			body = new String(
+					Base64.decodeBase64(response.getData().getBytes("UTF8")));
+			assertThat(body, notNullValue());
+			magic = body.substring(0, 9);
+			assertThat(magic, containsString("%PDF"));
+			writeToFile(Base64.decodeBase64(response.getData().getBytes("UTF8")),
+					filename);
+
+			PDF pdf = new PDF(
+					new File(System.getProperty("user.dir") + "/" + filename));
+			assertThat(pdf.text, containsString("The Free Encyclopedia"));
+			// NOTE: locale UTF8
+			assertThat(pdf.text, containsString("Русский"));
+			assertThat(pdf.text, containsString("Français"));
+			assertThat(pdf.encrypted, is(false));
+			assertThat(pdf.creator, is("Chromium"));
+			assertThat(pdf.numberOfPages, equalTo(2));
+			assertThat(pdf.getHeight(), equalTo(11.69));
+			assertThat(pdf.getWidth(), equalTo(8.26));
+		} catch (UnsupportedEncodingException e) {
+			System.err.println("Exception (ignored): " + e.toString());
+		} catch (IOException e) {
+			System.err.println("Exception (ignored): " + e.toString());
+		}
+
+	}
+	/*
+	 * response = chromeDevTools.send(new
+	 * Command<PrintToPDFResponse>("Page.printToPDF",
+	 * ImmutableMap.of("landscape", landscape),
+	 * ImmutableMap.of("displayHeaderFooter", displayHeaderFooter),
+	 * ImmutableMap.of("printBackground", printBackground),
+	 * ImmutableMap.of("scale", scale), ImmutableMap.of("paperWidth", null),
+	 * ImmutableMap.of("paperWidth", null), ImmutableMap.of("marginTop", null),
+	 * ImmutableMap.of("marginBottom", null), ImmutableMap.of("marginLeft",
+	 * null), ImmutableMap.of("marginRight", null),
+	 * ImmutableMap.of("pageRanges", null),
+	 * ImmutableMap.of("headerTemplate", null),
+	 * ImmutableMap.of("footerTemplate", null),
+	 * ImmutableMap.of("preferCSSPageSize", 0), ImmutableMap.of("transferMode",
+	 * transferMode), ConverterFunctions.map("data",
+	 * PrintToPDFResponse.class)));
+	 */
+
 
 	private void writeToFile(byte[] data, String fileName) {
 		try {
@@ -271,9 +336,20 @@ public class PrintToPDFDevToolsTest {
 		public final String producer;
 		public final String subject;
 		public final String title;
+
 		public final boolean encrypted;
 		public final boolean signed;
 		public final String signerName;
+		private double height;
+		private double width;
+
+		public double getHeight() {
+			return height;
+		}
+
+		public double getWidth() {
+			return width;
+		}
 
 		private PDF(String name, byte[] content) {
 			this(name, content, 1, Integer.MAX_VALUE);
@@ -296,7 +372,30 @@ public class PrintToPDFDevToolsTest {
 					this.producer = pdf.getDocumentInformation().getProducer();
 					this.subject = pdf.getDocumentInformation().getSubject();
 					this.title = pdf.getDocumentInformation().getTitle();
+
+					// find pdf page dimensions
+					// https://stackoverflow.com/questions/20904191/pdfbox-find-page-dimensions
+					float pageHeight = pdf.getPage(0).getMediaBox().getHeight() / 72;
+					float pageWidth = pdf.getPage(0).getMediaBox().getWidth() / 72;
+
+					// round down to 2 decimal places
+					// http://www.java2s.com/example/java-utility-method/decimal-round-index-1.html
+					NumberFormat format = NumberFormat.getInstance();
+					format.setGroupingUsed(false);
+					format.setMaximumFractionDigits(2);
+					try {
+						this.height = format.parse(format.format(pageHeight)).doubleValue();
+					} catch (ParseException e) {
+						this.height = pageHeight;
+					}
+					try {
+						this.width = format.parse(format.format(pageWidth)).doubleValue();
+					} catch (ParseException e) {
+						this.width = pageWidth;
+					}
+
 					this.encrypted = pdf.isEncrypted();
+					// https://pdfbox.apache.org/docs/2.0.3/javadocs/org/apache/pdfbox/pdmodel/common/PDRectangle.html
 
 					PDSignature signature = pdf.getLastSignatureDictionary();
 					this.signed = signature != null;
